Add files via upload
This is the first version of my attempt to create a LSTM RNN from scratch using data that was available on Kaggle.  I am not sure if it works just yet since I ran out of time to work on this.  I will hopefully work on this some more in the coming weeks.